refactor: 닉네임 표시를 위해 변경
